debarchive: set UID/GID to 0/0
Prior to this change, the ar file embeds the UID and GID of the invoking
user inside the deb archive.  After this change, the files in the
archive are owned by root.

In deb archives generated by standard Debian build tools (like
dpkg-buildpackage), the binary rule that generates the archive runs as
root (or with fakeroot).  Because the binary rule runs as root, the
contents of the ar file have UID and GID of 0.