Rewrite conffiles to ensure absolute paths
The man page for
[deb-conffiles](https://man7.org/linux/man-pages/man5/deb-conffiles.5.html)
states that conffiles should be absolute pathnames.

[Starting with dpkg
1.20.1](https://github.com/guillemj/dpkg/blob/68ab722604217d3ab836276acfc0ae1260b28f5f/debian/changelog#L393),
dpkg will fail to install packages that don't use absolute pathnames for
conffiles. Ubuntu 21.04 is a distro that is effected by this change.

This commit now ensures that conffile entries are absolute paths and
rewrites them if they are not.

This change has been tested on a Ubuntu 21.04 system.